fix rules connection

+++ src/server/endpoints/database/rules.py
@@ -4,7 +4,9 @@ from common import *
 
 # All rules for enforcing synchronization between shard replicas across servers
 # Simply do 'await rules(shard_id, term)' in each operation. No return value
+# MUST BE CALLED WITHIN A TRANSACTION CONTEXT
 async def rules(
+    conn,  # asyncpg connection object
     shard_id: str,
     valid_at: int
 ):
@@ -14,24 +16,25 @@ async def rules(
     """
 
     try:
+        # if not within transaction, then raise error
+        if not conn.is_in_transaction:
+            raise Exception('Not in transaction')
+        
         # Enforce rules by executing database operations
-        async with common.pool.acquire() as connection:
-            async with connection.transaction():
+        await conn.execute('''--sql
+            DELETE FROM StudT
+            WHERE (shard_id = $1::TEXT)
+                AND (created_at > $2::INTEGER
+                    OR (deleted_at IS NOT NULL
+                        AND deleted_at <= $2::INTEGER));
+            ''', shard_id, valid_at)
 
-                await connection.execute('''--sql
-                    DELETE FROM StudT
-                    WHERE (shard_id = $1::TEXT)
-                        AND (created_at > $2::INTEGER
-                            OR (deleted_at IS NOT NULL
-                                AND deleted_at <= $2::INTEGER));
-                    ''', shard_id, valid_at)
-
-                await connection.execute('''--sql
-                    UPDATE StudT
-                    SET deleted_at = NULL
-                    WHERE shard_id = $1::TEXT
-                        AND deleted_at > $2::INTEGER;
-                    ''', shard_id, valid_at)
+        await conn.execute('''--sql
+            UPDATE StudT
+            SET deleted_at = NULL
+            WHERE shard_id = $1::TEXT
+                AND deleted_at > $2::INTEGER;
+            ''', shard_id, valid_at)
 
     except Exception as e:
